A System to mange events on campus
The project will allow any generic university to keep track of the
attendance record of
students to academically oriented learning activities. This application
will keep track of
what events are happening at the school, as well as who attends them.
Generally,
event attendance has been estimated with the help of sign-in sheet , or
just not
handled at all in some cases. But with the database keeping track of
attendance,
popular events can be repeated, poorly attended ones can be updated or
removed,
and required participation can easily be tracked by professors.